profile edit getmethod

+++ src/main/java/com/baylogic/jdbc/CommonDAO.java
@@ -10,6 +10,6 @@ public interface CommonDAO {
 	public List<Doctors> getDoctorsBySearch(String searchType,  PGArrayGeneric searchTypeIds);
 	public List<Doctors> getDoctorsBySearch(String searchType,  Integer searchTypeId);
 	public List<Doctors> getDoctorsBySearch2(String searchType,  Long searchTypeId);
-	public List<DocSpecializations> getDoctorSpecializations(Integer userLoginId);
+	public List<DocSpecializations> getDoctorSpecializations(Long userLoginId);
 
 }
